Security: Fix CVE-2022-0860
If PAM is correctly configured and a user account is set to expired,
the expired user-account is still able to successfully log into
Cobbler in all places (Web UI, CLI & XMLRPC-API).

The same applies to user accounts with passwords set to be expired.

This patch is fixing this and checking that this behavior is now
correct via a reproducible test.